Split Inferencer.hs into utils, inferencer and simplifier
