+++ src/Inference/Inferencer.hs
@@ -5,7 +5,9 @@ import Inference.TypingEnvironment
 import Inference.Type
 import Inference.Substitutions
 import Inference.Errors
+import Inference.Simplifier
 import Inference.ConstraintSolver
+import Inference.InferencerUtils
 
 import AbsSyntax
 import PrintSyntax
@@ -18,7 +20,6 @@ import Data.Foldable
 
 import System.IO.Unsafe
 
-import Data.List (nub)
 import qualified Data.Map as Map
 import qualified Data.Set as Set
 
@@ -84,70 +85,6 @@ constraintsExpr env state ex = do
           sc = closeOver $ apply subst ty
 
 -- | Canonicalize and return the polymorphic toplevel type.
-closeOver :: Type -> Scheme
-closeOver = normalize . generalize Inference.TypingEnvironment.empty
-
--- | Lookup type in the environment
-lookupEnv :: Ident -> Infer Type
-lookupEnv x = do
-  (TypeEnv env) <- ask
-  case Map.lookup x env of
-      Nothing   ->  do payl <- errPayload
-                       throwError $ UnboundVariable payl x
-      Just s    ->  do t <- instantiate s
-                       return t
-
-letters :: [String]
-letters = [1..] >>= flip replicateM ['a'..'z']
-
-addExprAnnot :: Infer Expr -> Infer Expr
-addExprAnnot inExpr = do
-  s <- get
-  inferTraceTop <- return $ let InferState { inferTrace = inferTrace } = s in let (h:_) = inferTrace in h
-  e <- inExpr
-  return $ Annot inferTraceTop e
-
-markTrace :: (Show a, Print a) => a -> Infer ()
-markTrace a = do
-  s <- get
-  inferTrace <- return $ let InferState { inferTrace = inferTrace } = s in inferTrace
-  put s{ inferTrace = ([(printTree a)] ++ inferTrace) }
-  return ()
-
-unmarkTrace :: (Show a, Print a) => a -> Infer ()
-unmarkTrace a = do
-  s <- get
-  newTrace <- return $ let InferState { inferTrace = inferTrace } = s in drop 1 inferTrace
-  put s{ inferTrace = newTrace }
-  return ()
-
-errPayload :: Infer TypeErrorPayload
-errPayload = do
-  s <- get
-  lastTraceStr <- return $ let InferState { lastInferExpr = lastInferExpr } = s in lastInferExpr
-  return $ TypeErrorPayload lastTraceStr
-
-freshIdent :: Infer Ident
-freshIdent = do
-    s <- get
-    put s{count = count s + 1}
-    return $ Ident $ "__@$internal_variable__" ++ (letters !! count s) ++ "_"
-
-fresh :: Infer Type
-fresh = do
-    s <- get
-    put s{count = count s + 1}
-    return $ TVar $ TV (letters !! count s)
-
-instantiate ::  Scheme -> Infer Type
-instantiate (Forall as t) = do
-    as' <- mapM (const fresh) as
-    let s = Subst $ Map.fromList $ zip as as'
-    return $ apply s t
-
-generalize :: Env -> Type -> Scheme
-generalize env t  = Forall as t
-    where as = Set.toList $ ftv t `Set.difference` ftv env
 
 ops :: Binop -> Infer Type
 ops OpSemicolon = do
@@ -188,15 +125,6 @@ opsUni (OpTupleNth index len) = do
     tv <- fresh
     return $ ((TTuple tv tup), [tv] ++ tvs)) ((TTuple TUnit TUnit), []) (replicate len 0)
   return $ (tupleType) `TArr` (elsTypes !! index)
-
-withTypeAnnot :: TypeConstraint -> ComplexExpression -> Infer ComplexExpression
-withTypeAnnot TypeConstrEmpty e = return e
-withTypeAnnot (TypeConstrDef texpr) e = do
-  return $ ECChecked e texpr
-
-isRec :: LetRecKeyword -> Bool
-isRec LetRecYes = True
-isRec LetRecNo = False
 
 inferImplementation :: Implementation -> Infer (Env, Type, [AConstraint])
 inferImplementation ast@(IRoot cores) = do
@@ -366,20 +294,6 @@ resolveTypeExpressionRec fvs (TypeExprTuple fstEl restEls) = do
     return $ TTuple t acc) (TTuple TUnit TUnit) ([fstEl] ++ restEls)
   return tupleT
 
-getConstScheme :: Constant -> Scheme
-getConstScheme (CInt _) = Forall [] (TCon "Int")
-getConstScheme (CBool _) = Forall [] (TCon "Bool")
-getConstScheme (CString _) = Forall [] (TCon "String")
-
-getOperatorName :: OperatorAny -> String
-getOperatorName (OperatorAnyA (OperatorA name)) = name
-getOperatorName (OperatorAnyB (OperatorB name)) = name
-getOperatorName (OperatorAnyC (OperatorC name)) = name
-getOperatorName (OperatorAnyD (OperatorD name)) = name
-getOperatorName (OperatorAnyDS (OperatorDS)) = "*"
-getOperatorName (OperatorAnyE (OperatorE name)) = name
-getOperatorName (OperatorAnyF (OperatorF name)) = name
-
 resolveTypeExpression :: TypeExpression -> Infer Scheme
 resolveTypeExpression exp = do
   fvs <- getTypeExpressionFV exp
@@ -387,207 +301,9 @@ resolveTypeExpression exp = do
   fvsT <- return $ Map.elems fvs
   return $ Forall fvsT t
 
-simplifyPattern :: Bool -> SimplePattern -> Expr -> Expr -> Infer Expr
-simplifyPattern _ PatNone _ expr = addExprAnnot $ return expr
-simplifyPattern _ ast@(PatList (PList [])) letExpr expr = do
-  markTrace ast
-  scheme <- resolveTypeExpression (TypeExprSimple $ TypeSExprList $ TypeExprSimple $ TypeSExprAbstract $ TypeIdentAbstract "'a")
-  id <- freshIdent
-  unmarkTrace ast
-  addExprAnnot $ return $ Let id (Check letExpr scheme) expr
-simplifyPattern recMode (PatConstr (Ident nameStr) pat) letExpr expr = do
-  addExprAnnot $ simplifyPattern recMode pat (App (Var $ Ident $ nameStr ++ "_reverse") letExpr) expr
-simplifyPattern recMode ast@(PatList (PList list)) letExpr expr = do
-  markTrace ast
-  (simpl, _) <- foldrM (\(PListElement pat) (expr, n) -> do
-    patSimpl <- simplifyPattern recMode pat (UniOp OpListNth letExpr) expr
-    return (patSimpl, n+1)) (expr, 0) list
-  unmarkTrace ast
-  addExprAnnot $ return simpl
-simplifyPattern _ (PatConst const) letExpr expr = do
-  id <- freshIdent
-  addExprAnnot $ return $ Let id (Check letExpr $ getConstScheme const) expr
-simplifyPattern recMode (PatCheck name typeExpr) letExpr expr = do
-  scheme <- resolveTypeExpression typeExpr
-  addExprAnnot $ simplifyPattern recMode (PatIdent name) (Check letExpr scheme) expr
-simplifyPattern False (PatIdent name) letExpr expr =
-  addExprAnnot $ return $ Let name letExpr expr
-simplifyPattern True (PatIdent name) letExpr expr =
-  addExprAnnot $ return $ Let name (Fix (Lam name letExpr)) expr
-simplifyPattern recMode ast@(PatCons hPat tPat) letExpr expr = do
-  markTrace ast
-  hSimpl <- simplifyPattern recMode hPat (UniOp OpHead letExpr) expr
-  tSimpl <- simplifyPattern recMode tPat (UniOp OpTails letExpr) hSimpl
-  unmarkTrace ast
-  addExprAnnot $ return tSimpl
-simplifyPattern recMode ast@(PatTuple (PTuple el restEls)) letExpr expr = do
-  markTrace ast
-  tupleCount <- return $ 1 + length restEls
-  (tSimpl, _) <- foldrM (\(PTupleElement el) (expr, k) -> do
-    pSimpl <- simplifyPattern recMode el (UniOp (OpTupleNth k tupleCount) letExpr) expr
-    return (pSimpl, k+1)) (expr, 0) ([el] ++ restEls)
-  unmarkTrace ast
-  addExprAnnot $ return tSimpl
-
-simplifyComplexExpression :: ComplexExpression -> Infer Expr
-simplifyComplexExpression (ECTyped typeExpr) = do
-  scheme <- resolveTypeExpression typeExpr
-  addExprAnnot $ return $ Typed scheme
-simplifyComplexExpression ast@(ECFor varName initExpr dir endExpr bodyExpr) = do
-  markTrace ast
-  forCheckType <- return $ Forall [] $ (TCon "Int")
-  initSimpl <- simplifyComplexExpression initExpr
-  endSimpl <- simplifyComplexExpression endExpr
-  initSimplCheck <- return $ Check initSimpl forCheckType
-  endSimplCheck <- return $ Check endSimpl forCheckType
-  bodySimpl <- simplifyComplexExpression bodyExpr
-  unmarkTrace ast
-  addExprAnnot $ return $ Let varName (Op OpSame initSimplCheck endSimplCheck) bodySimpl
-simplifyComplexExpression ast@(ECWhile condExpr bodyExpr) = do
-  markTrace ast
-  whileCheckType <- return $ Forall [] $ (TCon "Bool")
-  condSimpl <- simplifyComplexExpression condExpr
-  bodySimpl <- simplifyComplexExpression bodyExpr
-  condSimplCheck <- return $ Check condSimpl whileCheckType
-  id <- freshIdent
-  unmarkTrace ast
-  addExprAnnot $ return $ Let id condSimplCheck bodySimpl
-  --bodyTypeExpr <- return $ TypeExprSimple $ TypeSExprEmpty
-  --simplifyComplexExpression $ ECLet LetRecNo (PatIdent $ Ident "x") [] (TypeConstrDef bodyTypeExpr) (ECLet LetRecNo (PatIdent $ varName) [] TypeConstrEmpty initExpr bodyExpr) $ ECExpr $ ExprVar $ Ident "x"
-
-simplifyComplexExpression ECExportEnv = addExprAnnot $ return Export
-simplifyComplexExpression (ECTuple tuple) = addExprAnnot $ simplifyTuple tuple
-simplifyComplexExpression (ECExpr expr) = addExprAnnot $ simplifyExpression expr
-simplifyComplexExpression ast@(ECIf cond exprThen exprElse) = do
-  markTrace ast
-  condSimpl <- simplifyComplexExpression cond
-  thenSimpl <- simplifyComplexExpression exprThen
-  elseSimpl <- simplifyComplexExpression exprElse
-  unmarkTrace ast
-  addExprAnnot $ return $ If condSimpl thenSimpl elseSimpl
-simplifyComplexExpression ast@(ECMatch expr _ clauses) = do
-  markTrace ast
-  clausesList <- foldrM (\(MatchClause pat clauseExpr) acc -> do
-    r <- return $ ListElement $ ECLet LetRecNo pat [] TypeConstrEmpty expr clauseExpr
-    return $ [r] ++ acc) [] clauses
-  r <- simplifyComplexExpression $ ECExpr $ ExprList $ DList clausesList
-  unmarkTrace ast
-  addExprAnnot $ return $ UniOp OpHead r
-
-simplifyComplexExpression ast@(ECFunction bPip matchClauses) = do
-  id <- freshIdent
-  markTrace ast
-  r <- simplifyComplexExpression $ ECFun (PatIdent id) [] $ ECMatch (ECExpr $ ExprVar id) bPip matchClauses
-  unmarkTrace ast
-  addExprAnnot $ return r
-simplifyComplexExpression ast@(ECFun pat argsPat expr) = do
-  id <- freshIdent
-  markTrace ast
-  r <- simplifyComplexExpression (ECLet LetRecNo (PatIdent id) ([pat] ++ argsPat) TypeConstrEmpty expr (ECExpr $ ExprVar id))
-  unmarkTrace ast
-  addExprAnnot $ return r
-simplifyComplexExpression (ECLetOperator recK opName patArgs letExpr expr) = do
-  --Ident $ getOperatorName opName
-  addExprAnnot $ simplifyComplexExpression (ECLet recK (PatIdent $ Ident $ getOperatorName opName) patArgs TypeConstrEmpty letExpr expr)
-simplifyComplexExpression ast@(ECLet recK pat [] typeAnnot letExpr expr) = do
-  markTrace ast
-  expr <- withTypeAnnot typeAnnot expr
-  letSimpl <- simplifyComplexExpression letExpr
-  exprSimpl <- simplifyComplexExpression expr
-  r <- simplifyPattern (isRec recK) pat letSimpl exprSimpl
-  unmarkTrace ast
-  addExprAnnot $ return r
-simplifyComplexExpression (ECChecked expr typeExpr) = do
-  simpl <- simplifyComplexExpression expr
-  s <- resolveTypeExpression typeExpr
-  addExprAnnot $ return $ Check simpl s
-simplifyComplexExpression ast@(ECLet recK pat argsPats typeAnnot letExpr expr) = do
-  markTrace ast
-  expr <- withTypeAnnot typeAnnot expr
-  letSimpl <- simplifyComplexExpression letExpr
-  exprSimpl <- simplifyComplexExpression expr
-  letSimplAcc <- foldrM (\pat expr -> do
-    id <- freshIdent
-    s <- simplifyPattern False pat (Var $ id) expr
-    return $ Lam id s) letSimpl argsPats
-  r <- simplifyPattern (isRec recK) pat letSimplAcc exprSimpl
-  unmarkTrace ast
-  addExprAnnot $ return r
-
-simplifyTuple :: DTuple -> Infer Expr
-simplifyTuple (DTuple firstElem elems) = do
-  elemsSimpl <- foldM (\expr (DTupleElement el) -> do
-    k <- simplifyExpression el
-    return $ Op OpTupleCons k expr) (UniOp OpEmptyTuple Skip) ([firstElem] ++ elems)
-  addExprAnnot $ return $ elemsSimpl
-
-simplifyList :: DList -> Infer Expr
-simplifyList (DList elems) = do
-  elemsSimpl <- foldM (\expr (ListElement el) -> do
-    k <- simplifyComplexExpression el
-    return $ Op OpCons k expr) (UniOp OpEmptyList Skip) elems
-  addExprAnnot $ return $ elemsSimpl
-
-simplifyExpression :: Expression -> Infer Expr
-simplifyExpression (ExprSemi expA expB) = do
-  simplA <- simplifyExpression expA
-  simplB <- simplifyExpression expB
-  addExprAnnot $ return $ Op OpSemicolon simplA simplB
-simplifyExpression (ExprOp opName) = do
-  addExprAnnot $ return $ Var $ Ident $ getOperatorName opName
-simplifyExpression (Expr1 expA (OperatorA name) expB) = do
-  simplA <- simplifyExpression expA
-  simplB <- simplifyExpression expB
-  addExprAnnot $ return $ Op (OpCustom name) simplA simplB
-simplifyExpression (Expr2 expA (OperatorB name) expB) = do
-  simplA <- simplifyExpression expA
-  simplB <- simplifyExpression expB
-  addExprAnnot $ return $ Op (OpCustom name) simplA simplB
-simplifyExpression (Expr3 expA (OperatorC name) expB) = do
-  simplA <- simplifyExpression expA
-  simplB <- simplifyExpression expB
-  addExprAnnot $ return $ Op (OpCustom name) simplA simplB
-simplifyExpression (Expr4 expA (OperatorD name) expB) = do
-  simplA <- simplifyExpression expA
-  simplB <- simplifyExpression expB
-  addExprAnnot $ return $ Op (OpCustom name) simplA simplB
-simplifyExpression (Expr4S expA OperatorDS expB) = do
-  simplA <- simplifyExpression expA
-  simplB <- simplifyExpression expB
-  addExprAnnot $ return $ Op (OpCustom "*") simplA simplB
-simplifyExpression (Expr5 expA (OperatorE name) expB) = do
-  simplA <- simplifyExpression expA
-  simplB <- simplifyExpression expB
-  addExprAnnot $ return $ Op (OpCustom name) simplA simplB
-simplifyExpression (Expr6 (OperatorF name) exp) = do
-  simpl <- simplifyExpression exp
-  addExprAnnot $ return $ UniOp (OpCustomUni name) simpl
-simplifyExpression (ExprVar name) = addExprAnnot $ return $ Var name
-simplifyExpression (ExprConst (CInt val)) = addExprAnnot $ return $ Lit $ LInt val
-simplifyExpression (ExprConst (CString val)) = addExprAnnot $ return $ Lit $ LString val
-simplifyExpression (ExprConst (CBool CBTrue)) = addExprAnnot $ return $ Lit $ LBool True
-simplifyExpression (ExprConst (CBool CBFalse)) = addExprAnnot $ return $ Lit $ LBool False
-simplifyExpression (ExprList list) = addExprAnnot $ simplifyList list
-simplifyExpression (ExprCompl expr) = addExprAnnot $ simplifyComplexExpression expr
-simplifyExpression ast@(ExprCall exp firstArg restArgs) = do
-  markTrace ast
-  fnExpr <- simplifyExpression exp
-  r <- foldlM (\simpl argExp -> do
-    argSimpl <- simplifySimpleExpression argExp
-    return $ App simpl argSimpl) (fnExpr) ([firstArg] ++ restArgs)
-  unmarkTrace ast
-  addExprAnnot $ return r
-
-simplifySimpleExpression :: SimpleExpression -> Infer Expr
-simplifySimpleExpression (ESOp opName) = simplifyExpression $ ExprOp opName
-simplifySimpleExpression (ESConst const) = simplifyExpression $ ExprConst const
-simplifySimpleExpression (ESIdent name) = simplifyExpression $ ExprVar name
-simplifySimpleExpression (ESExpr expr) = simplifyComplexExpression expr
-simplifySimpleExpression (ESList list) = simplifyList list
-
 inferComplexExpression :: ComplexExpression -> Infer (Type, [AConstraint])
 inferComplexExpression ast = do
-  tree <- simplifyComplexExpression ast
+  tree <- simplifyComplexExpression resolveTypeExpression ast
   infer tree
 
 inferE :: Expr -> Infer (Env, Type, [AConstraint])
@@ -699,28 +415,3 @@ infer expr = case expr of
     ac <- constraintAnnotList [(t1, typeBool), (t2, t3)]
     return (t2, c1 ++ c2 ++ c3 ++ ac)
 
-normalize :: Scheme -> Scheme
-normalize (Forall _ body) = Forall (map snd ord) (normtype body)
-  where
-    ord = zip (nub $ fv body) (map TV letters)
-
-    fv (TVar a)   = [a]
-    fv (TArr a b) = fv a ++ fv b
-    fv (TList a)  = fv a
-    fv (TTuple a b) = fv a ++ fv b
-    fv TUnit = []
-    fv (TExport _) = []
-    fv (TCon _)   = []
-    fv (TDep _ deps) = foldl (\acc el -> acc ++ (fv el)) [] deps
-
-    normtype TUnit = TUnit
-    normtype (TExport v) = (TExport v)
-    normtype (TArr a b) = TArr (normtype a) (normtype b)
-    normtype (TTuple a b) = TTuple (normtype a) (normtype b)
-    normtype (TList a) = TList (normtype a)
-    normtype (TCon a)   = TCon a
-    normtype (TDep name deps) = TDep name $ map normtype deps
-    normtype (TVar a)   =
-      case Prelude.lookup a ord of
-        Just x -> TVar x
-        Nothing -> error "type variable not in signature"
